Refactor: move mute v4 to .github/scripts/mute/, minimize create_new_muted_ya
- Create .github/scripts/mute/ package with logic, rules, data, apply, create_v4, run_parallel, compare, detect_manual, evaluate_pr
- Minimize create_new_muted_ya.py to legacy only: tests_monitor, hardcoded rules, no quarantine/graduation/pattern_rules/mute_decisions
- Move docs to mute/docs/
- Update workflows to use PYTHONPATH=.github/scripts python3 -m mute.run_parallel, mute.detect_manual, mute.compare
- Move unit tests to mute/tests/
- Delete obsolete files from tests/
