+++ .github/scripts/mute/__init__.py
@@ -1,0 +1,1 @@
+# Mute v4 package: logic, rules, data, apply, create_v4, run_parallel, compare, etc.
