+++ .github/scripts/mute/compare.py
@@ -1,18 +1,6 @@
 #!/usr/bin/env python3
 """
-Compare mute systems: legacy vs current (v4) vs v4_direct.
-
-Runs three modes:
-  - Legacy: create_new_muted_ya --legacy (uses tests_monitor, no quarantine)
-  - Current: create_new_muted_ya (uses tests_monitor, with quarantine)
-  - v4_direct: create_new_muted_ya_v4 (uses test_results directly, NO flaky_tests_window/tests_monitor)
-
-Legacy and Current require flaky_tests_history + tests_monitor. v4_direct does not.
-
-Usage:
-  compare_mute_systems.py --branch main [--build_type relwithdebinfo]
-  # For legacy/current: run flaky_tests_history.py and tests_monitor.py first
-  # v4_direct: only needs test_results (test_runs_column)
+Compare mute systems: legacy vs current vs v4_direct.
 """
 
 import argparse
@@ -29,89 +17,75 @@ def _load_lines(path):
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Compare mute systems: legacy vs current vs v4_direct")
+    parser = argparse.ArgumentParser(description="Compare mute systems")
     parser.add_argument("--branch", default="main")
     parser.add_argument("--build_type", default="relwithdebinfo")
     parser.add_argument("--muted_ya_file", default="base_muted_ya.txt")
     parser.add_argument("--quarantine_file", default="quarantine.txt")
     parser.add_argument("--output_dir", default="comparison")
-    parser.add_argument("--skip_legacy", action="store_true", help="Skip legacy (requires flaky+tests_monitor)")
-    parser.add_argument("--skip_current", action="store_true", help="Skip current (requires flaky+tests_monitor)")
-    parser.add_argument("--skip_v4_direct", action="store_true", help="Skip v4_direct (uses test_results only)")
+    parser.add_argument("--skip_legacy", action="store_true")
+    parser.add_argument("--run_current", action="store_true", help="Run current mode (legacy script is minimal, current=legacy now)")
+    parser.add_argument("--skip_v4_direct", action="store_true")
     args = parser.parse_args()
 
+    cwd = os.getcwd()
+
+    def _abs(p):
+        return os.path.join(cwd, p) if p and not os.path.isabs(p) else p
+
     script_dir = os.path.dirname(os.path.abspath(__file__))
-    create_script = os.path.join(script_dir, "create_new_muted_ya.py")
-    v4_direct_script = os.path.join(script_dir, "create_new_muted_ya_v4.py")
+    scripts_dir = os.path.dirname(script_dir)
+    create_script = os.path.join(scripts_dir, "tests", "create_new_muted_ya.py")
+    v4_script = os.path.join(script_dir, "create_v4.py")
     legacy_dir = os.path.join(args.output_dir, "legacy")
     current_dir = os.path.join(args.output_dir, "current")
     v4_direct_dir = os.path.join(args.output_dir, "v4_direct")
 
     os.makedirs(args.output_dir, exist_ok=True)
 
-    # Run legacy (needs flaky_tests_history + tests_monitor)
     if not args.skip_legacy:
         os.makedirs(legacy_dir, exist_ok=True)
         print("=== Running LEGACY (tests_monitor, no quarantine) ===")
         rc = subprocess.run(
-            [
-                sys.executable,
-                create_script,
-                "update_muted_ya",
-                "--branch", args.branch,
-                "--build_type", args.build_type,
-                "--muted_ya_file", args.muted_ya_file,
-                "--quarantine_file", args.quarantine_file,
-                "--output_folder", legacy_dir,
-            ] + ["--legacy"],
+            [sys.executable, create_script, "update_muted_ya",
+             "--branch", args.branch, "--build_type", args.build_type,
+             "--muted_ya_file", _abs(args.muted_ya_file),
+             "--quarantine_file", _abs(args.quarantine_file),
+             "--output_folder", legacy_dir, "--legacy"],
             capture_output=False,
         )
         if rc.returncode != 0:
-            print(f"Legacy run failed: {rc.returncode}")
             return rc.returncode
 
-    # Run current (needs flaky_tests_history + tests_monitor)
-    if not args.skip_current:
+    if args.run_current:
         os.makedirs(current_dir, exist_ok=True)
-        print("\n=== Running CURRENT (tests_monitor, with quarantine) ===")
+        print("\n=== Running CURRENT (same as legacy now) ===")
         rc = subprocess.run(
-            [
-                sys.executable,
-                create_script,
-                "update_muted_ya",
-                "--branch", args.branch,
-                "--build_type", args.build_type,
-                "--muted_ya_file", args.muted_ya_file,
-                "--quarantine_file", args.quarantine_file,
-                "--output_folder", current_dir,
-            ],
+            [sys.executable, create_script, "update_muted_ya",
+             "--branch", args.branch, "--build_type", args.build_type,
+             "--muted_ya_file", _abs(args.muted_ya_file),
+             "--quarantine_file", _abs(args.quarantine_file),
+             "--output_folder", current_dir, "--legacy"],
             capture_output=False,
         )
         if rc.returncode != 0:
-            print(f"Current run failed: {rc.returncode}")
             return rc.returncode
 
-    # Run v4_direct (NO flaky_tests_history, NO tests_monitor - read test_results only)
     if not args.skip_v4_direct:
         os.makedirs(v4_direct_dir, exist_ok=True)
-        print("\n=== Running v4_direct (test_results only, no flaky/monitor) ===")
+        print("\n=== Running v4_direct (test_results only) ===")
         rc = subprocess.run(
-            [
-                sys.executable,
-                v4_direct_script,
-                "--branch", args.branch,
-                "--build_type", args.build_type,
-                "--muted_ya_file", args.muted_ya_file,
-                "--quarantine_file", args.quarantine_file,
-                "--output_folder", v4_direct_dir,
-            ],
+            [sys.executable, "-m", "mute.create_v4",
+             "--branch", args.branch, "--build_type", args.build_type,
+             "--muted_ya_file", _abs(args.muted_ya_file),
+             "--quarantine_file", _abs(args.quarantine_file),
+             "--output_folder", v4_direct_dir],
             capture_output=False,
+            cwd=scripts_dir,
         )
         if rc.returncode != 0:
-            print(f"v4_direct run failed: {rc.returncode}")
             return rc.returncode
 
-    # Build comparison report
     mute_update = "mute_update"
     legacy_base = os.path.join(legacy_dir, mute_update)
     current_base = os.path.join(current_dir, mute_update)
@@ -120,23 +94,15 @@ def main():
     def load(name, base):
         return _load_lines(os.path.join(base, name)) if os.path.exists(base) else set()
 
-    report = []
-    report.append("# Mute System Comparison: Legacy vs Current vs v4_direct")
-    report.append(f"Branch: {args.branch}, Build type: {args.build_type}\n")
-    report.append("| Mode | Data source | flaky/monitor |")
-    report.append("|------|-------------|---------------|")
-    report.append("| Legacy | tests_monitor | required |")
-    report.append("| Current | tests_monitor | required |")
-    report.append("| v4_direct | test_results | **not used** |\n")
-
     modes = []
     if not args.skip_legacy:
         modes.append(("legacy", legacy_base))
-    if not args.skip_current:
+    if args.run_current:
         modes.append(("current", current_base))
     if not args.skip_v4_direct:
         modes.append(("v4_direct", v4_direct_base))
 
+    report = ["# Mute System Comparison", f"Branch: {args.branch}, Build type: {args.build_type}\n"]
     for label, filename in [
         ("to_mute", "to_mute.txt"),
         ("to_unmute", "to_unmute.txt"),
@@ -148,12 +114,10 @@ def main():
         for mode, vals in data.items():
             report.append(f"- {mode}: {len(vals)}")
         if len(modes) >= 2:
-            # Compare pairwise
             for i, (m1, b1) in enumerate(modes):
                 for m2, b2 in modes[i + 1:]:
                     s1, s2 = data[m1], data[m2]
-                    only_1 = s1 - s2
-                    only_2 = s2 - s1
+                    only_1, only_2 = s1 - s2, s2 - s1
                     if only_1 or only_2:
                         report.append(f"\n**{m1} vs {m2}:**")
                         if only_1:
@@ -174,12 +138,10 @@ def main():
 
     report_text = "\n".join(report)
     print("\n" + report_text)
-
     report_path = os.path.join(args.output_dir, "comparison_report.md")
     with open(report_path, "w") as f:
         f.write(report_text)
     print(f"\nReport saved to {report_path}")
-
     return 0
