+++ .github/scripts/mute/tests/__init__.py
@@ -1,0 +1,1 @@
+# Mute package tests
