+++ .github/workflows/update_muted_ya.yml
@@ -138,7 +138,7 @@ jobs:
       - name: Detect manual unmutes and update quarantine
         run: |
           if [ -f mute_state/previous_base.txt ] && [ -f mute_state/our_to_unmute.txt ]; then
-            python3 .github/scripts/tests/detect_manual_unmutes.py \
+            PYTHONPATH=.github/scripts python3 -m mute.detect_manual \
               --previous_base mute_state/previous_base.txt \
               --current_base base_muted_ya.txt \
               --to_unmute mute_state/our_to_unmute.txt \
@@ -158,7 +158,7 @@ jobs:
 
       - name: Run parallel mute update (legacy + v4)
         run: |
-          python3 .github/scripts/tests/run_parallel_mute_update.py \
+          PYTHONPATH=.github/scripts python3 -m mute.run_parallel \
             --branch=${{ env.BASE_BRANCH }} \
             --muted_ya_file=base_muted_ya.txt \
             --mute_muted_ya_file=base_mute_muted_ya.txt \
